Change pocl_binary interface
- removes the structs from header, they're really implementation details
- replace the binary interface by 7 new functions
