+++ lib/CL/pocl_binary.h
@@ -34,67 +34,36 @@ extern "C" {
 #pragma GCC visibility push(hidden)
 #endif
 
-/* pocl binary identifier */
-#define POCLCC_STRING_ID "poclbin"
-#define POCLCC_STRING_ID_LENGTH 7
-#define POCLCC_VERSION 1  
 
-/* pocl binary structures */
-typedef struct pocl_binary_kernel_s {
-  uint32_t sizeof_kernel_name;
-  uint32_t num_args;
-  uint32_t num_locals;
-  uint32_t sizeof_binary;
-  char *kernel_name;
-  struct pocl_argument *dyn_arguments;
-  struct pocl_argument_info *arg_info;
-  unsigned char *binary;
-} pocl_binary_kernel;
+/* check if buffer is pocl binary format, returns 1 if true */
+int pocl_binary_check_binary(cl_device_id device, unsigned char *binary);
 
-typedef struct pocl_binary_s {
-  char endian;
-  char pocl_id[POCLCC_STRING_ID_LENGTH];
-  uint64_t device_id;
-  uint32_t version;
-  uint32_t num_kernels;
-  pocl_binary_kernel *kernels;
-} pocl_binary;
 
-/* free internal structures */
-void pocl_binary_free_kernel(pocl_binary_kernel *kernel);
 
-/* check binary format validity */
-int pocl_binary_check_binary(cl_device_id device, pocl_binary *binary);
+/* returns the size of pocl_binaries[device_i] for allocation */
+size_t pocl_binary_sizeof_binary(cl_program program, unsigned device_i);
 
-/* get size of struct (serialized) */
-size_t pocl_binary_sizeof_binary(pocl_binary *binary);
+/* unpacks the content of program->pocl_binaries[device_i] into pocl cache */
+cl_int pocl_binary_untar(cl_program program, unsigned device_i);
 
-/* serialization/deserialization of pocl binary */
-int pocl_binary_serialize_binary(unsigned char *buffer, size_t sizeof_buffer,
-                                 pocl_binary *binary);
-int pocl_binary_deserialize_binary(pocl_binary *binary,
-                                   unsigned char *buffer, size_t sizeof_buffer);
+/* pocl cache -> program->pocl_binaries[device_i] */
+cl_int pocl_binary_serialize(cl_program program, unsigned device_i, size_t *size);
 
-/* initialize cl_kernel data from a pocl_binary_kernel specify 
-   with kernel_name and cl_device_id */
-cl_int pocl_binary_add_clkernel_data(unsigned char **binaries, int num_devices,
-                                   const char *kernel_name, cl_kernel kernel, 
-                                   cl_device_id device);
 
-/* look for a binary with kernel_name and cl_device_id */
-int pocl_binary_search_kernel_binary(unsigned char **binaries, int num_devices,
-                                     cl_device_id device, const char *kernel_name, 
-                                     unsigned char **binary, int *binary_size);
 
-/* initialize structures */
-void pocl_binary_init_binary(pocl_binary *binary, cl_device_id device, 
-                             int num_kernels, pocl_binary_kernel *kernels);
-int pocl_binary_init_kernel(pocl_binary_kernel *kernel, 
-                             char *kernel_name, int sizeof_kernel_name,
-                             unsigned char *binary, int sizeof_binary, 
-                             int num_args, int num_locals,
-                             struct pocl_argument *dyn_arguments, 
-                             struct pocl_argument_info *arg_info);
+/* returns the number of kernels without unpacking the binary */
+cl_uint pocl_binary_get_kernel_count(unsigned char *binary);
+
+/* returns the kernel names without unpacking the binary */
+cl_int pocl_binary_get_kernel_names(unsigned char *binary,
+                                    char **kernel_names,
+                                    size_t num_kernels);
+
+/* sets up cl_kernel's metadata, without unpacking the binary in pocl kcache */
+cl_int pocl_binary_get_kernel_metadata(unsigned char *binary,
+                                const char *kernel_name,
+                                cl_kernel kernel, cl_device_id device);
+
 
 #ifdef __GNUC__
 #pragma GCC visibility pop
